+++ tests/functional/adapter/incremental/test_incremental_strategies.py
@@ -33,7 +33,7 @@ class AppendBase(IncrementalBase):
             "append_model.sql": fixtures.base_model,
         }
 
-    def test_append(self, project):
+    def test_incremental(self, project):
         self.seed_and_run_twice()
         util.check_relations_equal(project.adapter, ["append_model", "append_expected"])
 
@@ -68,6 +68,7 @@ class TestAppendParquetHive(AppendBase):
         }
 
 
+@pytest.mark.skip_profile("databricks_uc_sql_endpoint")
 class InsertOverwriteBase(IncrementalBase):
     @pytest.fixture(scope="class")
     def seeds(self):
@@ -85,15 +86,17 @@ class InsertOverwriteBase(IncrementalBase):
             "overwrite_model.sql": fixtures.base_model,
         }
 
-    def test_append(self, project):
+    def test_incremental(self, project):
         self.seed_and_run_twice()
         util.check_relations_equal(project.adapter, ["overwrite_model", "overwrite_expected"])
 
 
+@pytest.mark.skip_profile("databricks_uc_sql_endpoint")
 class TestInsertOverwriteDelta(InsertOverwriteBase):
     pass
 
 
+@pytest.mark.skip_profile("databricks_uc_sql_endpoint")
 class TestInsertOverwriteWithPartitionsDelta(InsertOverwriteBase):
     @pytest.fixture(scope="class")
     def project_config_update(self):
@@ -104,8 +107,18 @@ class TestInsertOverwriteWithPartitionsDelta(InsertOverwriteBase):
             }
         }
 
+    @pytest.fixture(scope="class")
+    def seeds(self):
+        return {
+            "upsert_expected.csv": fixtures.upsert_expected,
+        }
 
-@pytest.mark.skip_profile("databricks_uc_cluster", "databricks_cluster")
+    def test_incremental(self, project):
+        self.seed_and_run_twice()
+        util.check_relations_equal(project.adapter, ["overwrite_model", "upsert_expected"])
+
+
+@pytest.mark.skip("This test is not repeatable due to external location")
 class TestInsertOverwriteParquet(InsertOverwriteBase):
     @pytest.fixture(scope="class")
     def project_config_update(self):
@@ -119,7 +132,7 @@ class TestInsertOverwriteParquet(InsertOverwriteBase):
         }
 
 
-@pytest.mark.skip_profile("databricks_uc_cluster", "databricks_cluster")
+@pytest.mark.skip("This test is not repeatable due to external location")
 class TestInsertOverwriteWithPartitionsParquet(InsertOverwriteBase):
     @pytest.fixture(scope="class")
     def project_config_update(self):